nrinow/requirements.txt
# NRI NOW Scraper Requirements
undetected-chromedriver>=3.5.0
selenium>=4.0.0
beautifulsoup4>=4.11.0
requests>=2.25.0
newspaper4k>=0.9.0
